ml-service/requirements.txt
flask==2.3.0
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.2
python-dotenv==1.0.0
cors==1.0.1
